No separate storage sets
